apps/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
redis>=5.0.1
python-dotenv>=1.0.0
pydantic>=2.6.0
email-validator>=2.0.0
openai>=1.3.0
chromadb>=0.4.0
langchain-community>=0.0.20
numpy>=1.24.0
faiss-cpu>=1.7.4
pymongo>=4.6.0
beautifulsoup4>=4.12.0
rapidfuzz>=3.0.0
sentence-transformers>=2.2.0
apscheduler>=3.10.4
pillow>=10.0.0
requests>=2.31.0
playwright==1.54.0